Fold saved.save router logic into the saveArticle service (#956)
saved.save in the router was ~540 lines duplicating (and diverging from)
the saveArticle service used by MCP: the service enforced the article size
limit and did Markdown-frontmatter extraction while the router did neither,
and the service ran Readability synchronously on the request thread while
the router used the worker pool.

saveArticle now owns the full flow with the router-only capabilities as
options: provided HTML (bookmarklet), refetch/force with the content-quality
guard, and the interactive private-Google-Docs OAuth flow (googleDocsAuth,
public docs already came via the plugin). It returns an outcome
(created/updated/existing) so the router can pick the right counts query,
runs Readability in the worker pool for all callers, and enforces
maxSavedArticleSizeBytes on every acquisition path including provided HTML
(also capped by Zod on the router input). saved.save and saved.delete are
now thin wrappers over the service, and the verbatim extractMetadata copy
is gone.
